Define ssize_t as intptr_t to avoid including BaseTsd.h for MSVC
MSDN states that SSIZE_T is defined as LONG_PTR, and LONG_PTR is defined as
long (32-bit) when building for x86 and __int64 (64-bit) when building for
x64, which basically means that it's signed and its size matches that of a
pointer, which in turn is the definition of intptr_t. The latter is
provided by stdint.h which we're already including.